src/ecommerces_mysql/requirements.txt
rq
requests
parse
pathlib

mysql-connector-python
pymongo